chore: feature event formatter format style
