+++ src/lib/addons/new-relic.ts
@@ -12,7 +12,6 @@ import {
 import {
     type FeatureEventFormatter,
     FeatureEventFormatterMd,
-    LinkStyle,
 } from './feature-event-formatter-md';
 import { gzip } from 'node:zlib';
 import { promisify } from 'util';
@@ -45,10 +44,9 @@ export default class NewRelicAddon extends Addon {
 
     constructor(config: IAddonConfig) {
         super(definition, config);
-        this.msgFormatter = new FeatureEventFormatterMd(
-            config.unleashUrl,
-            LinkStyle.MD,
-        );
+        this.msgFormatter = new FeatureEventFormatterMd({
+            unleashUrl: config.unleashUrl,
+        });
         this.flagResolver = config.flagResolver;
     }
